chore: instala ts-node
